@@ -4,13 +4,15 @@ namespace App\Models;
 
 use App\User;
 use App\Models\Review;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Product extends Model
 {
-    protected $fillable = [
-        'name','image','detail','stock','discount','price'
-    ];
+    use HasFactory;
+
+    protected $fillable = ['name','image','detail','stock','discount','price'];
+
     public function reviews(){
         return $this->hasMany(Review::class);
     }
